refactor(client): simplify code — deduplicate utils, extract SSE parser helper (-96 lines)
