+++ packages/client/src/sse.ts
@@ -28,6 +28,35 @@ export interface SSEControlEvent {
 export type SSEEvent = SSEDataEvent | SSEControlEvent
 
 /**
+ * Parse a control event JSON string into an SSEControlEvent.
+ * Throws DurableStreamError on invalid JSON.
+ */
+function parseControlEvent(dataStr: string): SSEControlEvent {
+  try {
+    const control = JSON.parse(dataStr) as {
+      streamNextOffset: Offset
+      streamCursor?: string
+      upToDate?: boolean
+      streamClosed?: boolean
+    }
+    return {
+      type: `control`,
+      streamNextOffset: control.streamNextOffset,
+      streamCursor: control.streamCursor,
+      upToDate: control.upToDate,
+      streamClosed: control.streamClosed,
+    }
+  } catch (err) {
+    const preview =
+      dataStr.length > 100 ? dataStr.slice(0, 100) + `...` : dataStr
+    throw new DurableStreamError(
+      `Failed to parse SSE control event: ${err instanceof Error ? err.message : String(err)}. Data: ${preview}`,
+      `PARSE_ERROR`
+    )
+  }
+}
+
+/**
  * Parse SSE events from a ReadableStream<Uint8Array>.
  * Yields parsed events as they arrive.
  */
@@ -69,29 +98,7 @@ export async function* parseSSEStream(
             if (currentEvent.type === `data`) {
               yield { type: `data`, data: dataStr }
             } else if (currentEvent.type === `control`) {
-              try {
-                const control = JSON.parse(dataStr) as {
-                  streamNextOffset: Offset
-                  streamCursor?: string
-                  upToDate?: boolean
-                  streamClosed?: boolean
-                }
-                yield {
-                  type: `control`,
-                  streamNextOffset: control.streamNextOffset,
-                  streamCursor: control.streamCursor,
-                  upToDate: control.upToDate,
-                  streamClosed: control.streamClosed,
-                }
-              } catch (err) {
-                // Control events contain critical offset data - don't silently ignore
-                const preview =
-                  dataStr.length > 100 ? dataStr.slice(0, 100) + `...` : dataStr
-                throw new DurableStreamError(
-                  `Failed to parse SSE control event: ${err instanceof Error ? err.message : String(err)}. Data: ${preview}`,
-                  `PARSE_ERROR`
-                )
-              }
+              yield parseControlEvent(dataStr)
             }
             // Unknown event types are silently skipped per protocol
           }
@@ -125,28 +132,7 @@ export async function* parseSSEStream(
       if (currentEvent.type === `data`) {
         yield { type: `data`, data: dataStr }
       } else if (currentEvent.type === `control`) {
-        try {
-          const control = JSON.parse(dataStr) as {
-            streamNextOffset: Offset
-            streamCursor?: string
-            upToDate?: boolean
-            streamClosed?: boolean
-          }
-          yield {
-            type: `control`,
-            streamNextOffset: control.streamNextOffset,
-            streamCursor: control.streamCursor,
-            upToDate: control.upToDate,
-            streamClosed: control.streamClosed,
-          }
-        } catch (err) {
-          const preview =
-            dataStr.length > 100 ? dataStr.slice(0, 100) + `...` : dataStr
-          throw new DurableStreamError(
-            `Failed to parse SSE control event: ${err instanceof Error ? err.message : String(err)}. Data: ${preview}`,
-            `PARSE_ERROR`
-          )
-        }
+        yield parseControlEvent(dataStr)
       }
     }
   } finally {
